refactor: 移除 BytecodeBuilder 抽象基类并新增泛型 Emit
- 删除 BytecodeBuilder 抽象基类与工厂方法 Create，改用具体类直接构造
- 新增 Emit<T>(OpCode, T) 泛型方法，按枚举底层大小（byte/short/int）编码参数
- 调用点改为直接传入枚举，移除 (int) 显式强转

+++ PySharp/Compilation/Bytecodes/Emitter.Expr.cs
@@ -266,7 +266,7 @@ partial class Emitter
     {
         LoadExpr(node.Left);
         LoadExpr(node.Right);
-        Builder.Emit(OpCode.BinaryOp, (int)node.Operator);
+        Builder.Emit(OpCode.BinaryOp, node.Operator);
     }
 
     private void EmitUnaryOp(UnaryOpNode node)
@@ -279,7 +279,7 @@ partial class Emitter
         }
         else
         {
-            Builder.Emit(OpCode._UnaryOp, (int)node.Op);
+            Builder.Emit(OpCode._UnaryOp, node.Op);
         }
     }
 
@@ -336,7 +336,7 @@ partial class Emitter
             else if (op is CmpopType.In or CmpopType.NotIn)
                 Builder.Emit(OpCode.ContainsOp, op is CmpopType.In ? 0 : 1);
             else
-                Builder.Emit(OpCode.CompareOp, (int)op);
+                Builder.Emit(OpCode.CompareOp, op);
         }
     }
